Added vertical movement via keyboard. Added horizontal movement as well. Pieces no longer stack onto each other. Added a game over screen and basic logic

+++ Tetris/Tetris/TetroN.cpp
@@ -13,7 +13,7 @@ TetroN::~TetroN()
 
 int TetroN::getPiece(int pX, int pY)
 {
-	return (piece[static_cast<int>(actualRotation)][pX][pY]);
+	return (piece[actualRotation][pX][pY]);
 }
 
 void TetroN::rotateRight()
